+++ CHANGELOG.md
@@ -7,8 +7,8 @@
     - `toMapOfCanonicalKeys`: creates a `Map<C,V>` (with the canonicalized keys).
 - Fixes bugs in `ListSlice.slice` and `ListExtensions.slice`.
 - Adds `shuffled` to `IterableExtension`.
+- Shuffle `IterableExtension.sample` results.
 - Update to `package:lints` 2.0.1.
-- Better randomization of `IterableExtension.sample` results.
 
 ## 1.17.2