+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.19.0-wip
+
+- Adds `shuffled` to `IterableExtension`.
+
 ## 1.18.0
 
 - `CanonicalizedMap`:
@@ -6,7 +10,6 @@
     - `toMap`: creates a `Map<K,V>` (with the original key values).
     - `toMapOfCanonicalKeys`: creates a `Map<C,V>` (with the canonicalized keys).
 - Fixes bugs in `ListSlice.slice` and `ListExtensions.slice`.
-- Adds `shuffled` to `IterableExtension`.
 - Update to `package:lints` 2.0.1.
 
 ## 1.17.2